refactor budget detail page to fix static params build error
